requirements.txt
asgiref==3.8.1
boto3==1.37.38
botocore==1.37.38
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==4.2.20
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-seed==0.3.1
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.1.0
environ==1.0
gitdb==4.0.12
GitPython==3.1.41
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
moto==5.1.4
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
responses==0.25.7
s3transfer==0.11.5
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
xmltodict==0.14.2
<<<<<<< HEAD
smmap==5.0.2
sqlparse==0.5.3
typing_extensions==4.13.0
=======
black==24.4.2
flake8==7.0.0
>>>>>>> 2853bf3805e39ed850dac0c989affcba4e0192cf
